requirements.txt
asgiref==3.8.1
chardet==5.2.0
DateTimeRange==2.3.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
install==1.3.5
mbstrdecoder==1.1.3
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typepy==1.3.2
typing_extensions==4.11.0
tzdata==2024.1
